Update post holiday
Fixed function that handles how the data is sent when a show is selected for full screen view. The whole object is passed now to allow for more information to be added in future. Added back button to selected screen. If filter results in no shoes then condition statement will be rendered to inform user of this.